Created Match class & altered Athlete class
Created the Match class.

The Athlete class now has java documentation for parameters and returns for getter and setter methods.

The setStamina method has been altered so that the stamina of the Athlete cannot be less than 0, and if it becomes 0 then the setIsInjured method is called, setting the isInjured property to True.